requirements.txt
scikit-image
numba
pandas
termcolor
h5py
tqdm
matplotlib
imageio
ipykernel
torchinfo
hydra-core
wandb
gitignore-parser
opencv-python-headless
hdf5plugin
nbformat
nbstripout
rerun-sdk
h5pickle
mkl-service